addeed count reviews

+++ myreviewapplication/Models/ViewModel/review.cs
@@ -6,6 +6,6 @@
         public string  CommentText { get; set; }
         public string UserName { get; set; } 
         public Guid CommentId { get; set; } 
-        public int rating { get; set; }
+        public double rating { get; set; }
     }
 }
